chore: rename partials to components
