+++ components/common.js
